# content/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetView": {
        "additionalProperties": false,
        "properties": {
          "content_ref": {
            "title": "Content Ref",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "style_pack_version": {
            "title": "Style Pack Version",
            "type": "string"
          },
          "subject_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "subject_visual_version": {
            "title": "Subject Visual Version",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "world_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "World Id"
          }
        },
        "required": [
          "id",
          "kind",
          "content_ref",
          "mime",
          "width",
          "height",
          "style_pack_version",
          "subject_visual_version",
          "status",
          "version"
        ],
        "title": "AssetView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "worldsim",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/library/assets": {
      "get": {
        "description": "Unscoped (preset-pickable) ready assets; world-bound art stays world-scoped.",
        "operationId": "list_library_assets_api_v1_library_assets_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "portrait",
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetView"
                  },
                  "title": "Response List Library Assets Api V1 Library Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Library Assets",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/assets/{asset_id}/bytes": {
      "get": {
        "description": "Serve unscoped bytes only; world-bound assets 404 here by design.",
        "operationId": "read_library_asset_bytes_api_v1_library_assets__asset_id__bytes_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Library Asset Bytes",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/import/apply": {
      "post": {
        "description": "Explicit atomic apply after a validate preview.",
        "operationId": "apply_import_api_v1_library_import_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Import",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/import/validate": {
      "post": {
        "description": "Preview validation only; writes nothing.",
        "operationId": "validate_import_api_v1_library_import_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Import Api V1 Library Import Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Import",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets": {
      "get": {
        "operationId": "list_presets_api_v1_library_presets_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PresetSummary"
                  },
                  "title": "Response List Presets Api V1 Library Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Presets",
        "tags": [
          "library"
        ]
      },
      "post": {
        "operationId": "create_preset_api_v1_library_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Preset",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}": {
      "get": {
        "operationId": "read_preset_api_v1_library_presets__preset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Preset",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}/archive": {
      "post": {
        "operationId": "archive_preset_api_v1_library_presets__preset_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Preset",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}/assets": {
      "get": {
        "description": "Resolve a preset revision's asset references with explicit miss warnings.",
        "operationId": "preset_assets_api_v1_library_presets__preset_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preset Assets Api V1 Library Presets  Preset Id  Assets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preset Assets",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}/duplicate": {
      "post": {
        "description": "Copy the current revision into a new editable preset; built-ins stay put.",
        "operationId": "duplicate_preset_api_v1_library_presets__preset_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Preset",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}/editor-drafts": {
      "post": {
        "description": "Open the durable draft for a base revision; same base replays it.",
        "operationId": "open_editor_draft_api_v1_library_presets__preset_id__editor_drafts_post",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorDraftOpenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorDraftView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Editor Draft",
        "tags": [
          "library"
        ]
      }
    },
    "/api/v1/library/presets/{preset_id}/editor-drafts/current": {
      "get": {
        "operationId": "read_editor_draft_api_v1_library_presets__preset_id__editor_drafts_current_get",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorDraftView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Editor Draft",
        "tags": [
          "library"
        ]
      }
    },
...
    "/api/v1/library/presets/{preset_id}/editor-drafts/{draft_id}/complete": {
      "post": {
        "description": "Retire a draft after its publication; refuses newer unpublished edits.",
        "operationId": "complete_editor_draft_api_v1_library_presets__preset_id__editor_drafts__draft_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preset Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorDraftCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Complete Editor Draft Api V1 Library Presets  Preset Id  Editor Drafts  Draft Id  Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Editor Draft",
        "tags": [
          "library"
        ]
      }
    },
...
  }
}